refactor: extract styles
