+++ src/components/Chart/Chart.tsx
@@ -10,71 +10,12 @@ import * as React from 'react';
 import { cn } from '@douglasneuroinformatics/libui/utils';
 import * as RechartsPrimitive from 'recharts';
 
-import { ChartContext } from '@/context/ChartContext';
 import { useChart } from '@/hooks/useChart';
 
+import { ChartContainer } from './ChartContainer';
+import { ChartStyle } from './ChartStyle';
+
 import type { ChartConfig } from './types';
-
-// Format: { THEME_NAME: CSS_SELECTOR }
-const THEMES = { dark: '.dark', light: '' } as const;
-
-const ChartContainer = React.forwardRef<
-  HTMLDivElement,
-  {
-    children: React.ComponentProps<typeof RechartsPrimitive.ResponsiveContainer>['children'];
-    config: ChartConfig;
-  } & React.ComponentProps<'div'>
->(({ children, className, config, id, ...props }, ref) => {
-  const uniqueId = React.useId();
-  const chartId = `chart-${id || uniqueId.replace(/:/g, '')}`;
-
-  return (
-    <ChartContext.Provider value={{ config }}>
-      <div
-        className={cn(
-          "[&_.recharts-cartesian-grid_line[stroke='#ccc']]:stroke-border/50 flex aspect-video justify-center text-xs [&_.recharts-cartesian-axis-tick_text]:fill-muted-foreground [&_.recharts-curve.recharts-tooltip-cursor]:stroke-border [&_.recharts-dot[stroke='#fff']]:stroke-transparent [&_.recharts-layer]:outline-none [&_.recharts-polar-grid_[stroke='#ccc']]:stroke-border [&_.recharts-radial-bar-background-sector]:fill-muted [&_.recharts-rectangle.recharts-tooltip-cursor]:fill-muted [&_.recharts-reference-line_[stroke='#ccc']]:stroke-border [&_.recharts-sector[stroke='#fff']]:stroke-transparent [&_.recharts-sector]:outline-none [&_.recharts-surface]:outline-none",
-          className
-        )}
-        data-chart={chartId}
-        ref={ref}
-        {...props}
-      >
-        <ChartStyle config={config} id={chartId} />
-        <RechartsPrimitive.ResponsiveContainer>{children}</RechartsPrimitive.ResponsiveContainer>
-      </div>
-    </ChartContext.Provider>
-  );
-});
-ChartContainer.displayName = 'Chart';
-
-const ChartStyle = ({ config, id }: { config: ChartConfig; id: string }) => {
-  const colorConfig = Object.entries(config).filter(([_, config]) => config.theme || config.color);
-
-  if (!colorConfig.length) {
-    return null;
-  }
-
-  return (
-    <style
-      dangerouslySetInnerHTML={{
-        __html: Object.entries(THEMES)
-          .map(
-            ([theme, prefix]) => `
-${prefix} [data-chart=${id}] {
-${colorConfig
-  .map(([key, itemConfig]) => {
-    const color = itemConfig.theme?.[theme as keyof typeof itemConfig.theme] || itemConfig.color;
-    return color ? `  --color-${key}: ${color};` : null;
-  })
-  .join('\n')}
-}
-`
-          )
-          .join('\n')
-      }}
-    />
-  );
-};
 
 const ChartTooltip = RechartsPrimitive.Tooltip;
